+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `rivet exec` command to run arbitrary commands with `RIVET_API_ENDPOINT` and `RIVET_TOKEN` environment variables
 - `rivet run` command to run scripts from the `scripts` portion of `rivet.yaml` with `RIVET_API_ENDPOINT` and `RIVET_TOKEN` environment variables
+- `RIVET_API_ENDPOINT` arg is passed to `docker build` by default
 
 ## [v1.0.0-rc.1] - 2023-12-24